🧪 Improve codeium configuration.
This is also quite careful not to read work's workspaces, and
additionally it will only be enabled further when a file exists in the
cwd called .codeium-enabled. I would have preferred to do this with
os.getenv(), but that can be a bit of a pain for direct launches in
Wezterm since it doesn't get the current context.